+++ changelog.md
@@ -2,9 +2,9 @@
 
 ## [Unreleased][unreleased]
 
+## [0.0.8] - 2018-09-05
 ### Added
 - Added support for nosql(couchdb)
-
 ### Removed
 - Removed default configs for database